ai_services/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-core
langchain-ollama
langchain-chroma
langchain-text-splitters
chromadb
pypdf
python-dotenv
python-multipart
